+++ toys-core/CHANGELOG.md
@@ -5,6 +5,7 @@
 # CHANGED: Cli#add_search_path_hierarchy changed the behavior of the base/terminate param
 * ADDED: Ability to define named templates in Toys files
 * ADDED: Ability to disable argument parsing
+* ADDED: Exec#exec_tool that supports all the stream redirects
 
 ### 0.3.8 / 2018-06-10